clears up the ViewModel
